% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisbinom.R
\name{dpoisbinom}
\alias{dpoisbinom}
\title{Probability Mass Function for the Poisson Binomial Distribution}
\usage{
dpoisbinom(x, probs, log = 0)
}
\arguments{
\item{x}{Integer scalar that contains the number of successful trials from
the Poisson Binomial process.}

\item{probs}{Numeric vector containing the probability of success of each of
the independent trails that comprise the Poisson Binomial distribution.}

\item{log}{Logical scalar indicating whether the log of the probability mass
is required.}
}
\value{
A numeric scalar containing the probability mass (or the log of the
probability mass if \code{log = TRUE}) of obtaining exactly \code{x}
successful trials.
}
\description{
Function that provides the probability mass function of the Poisson binomial
distribution.  This function is encapsulated in a way that it can extend the
functionality of the NIMBLE modelling framework.
}
\details{
The calculation of the probability masses is performed according to
the discrete Fourier transform of the Poisson binomial characteristic
function as described by Hong (2013).
}
\references{
Hong (2013) On computing the distribution function for the
Poisson binomial distribution \emph{Computational Statistics and Data Analysis}
59, 41-51.
}
\seealso{
\code{\link[nimble]{nimbleExternalCall}} \code{\link{poisbinom_massvec}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
