% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getfaaimeDistributions}
\alias{getfaaimeDistributions}
\title{Retrieve List of NIMBLE Distributions Defined in faaime}
\usage{
getfaaimeDistributions()
}
\value{
A named list with each element containing the distribution
specification required for the \code{\link[nimble]{registerDistributions}}
function.
}
\description{
Function to rertrieve a list of the NIMBLE distributions defined in the
faaime package.
}
\seealso{
\code{\link[nimble]{registerDistributions}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
