% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisbinom.R
\name{nimatan2}
\alias{nimatan2}
\title{NIMBLE Implementation of arctan2 function}
\usage{
nimatan2(iny, inx)
}
\arguments{
\item{iny}{Numeric scalar}

\item{inx}{Numeric scalar}
}
\value{
A numeric scalar returning the angle (in radians) between the x-axis
and the vector from the origin to (\code{inx}, \code{iny}).
}
\description{
A function for NIMBLE to allow the calculation of the arctan2 function. This
function has not been vectorised.
}
\seealso{
\code{\link[nimble]{nimbleExternalCall}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
