% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{nimble_fft}
\alias{nimble_fft}
\title{NIMBLE FFT External Call}
\usage{
nimble_fft(pReal, pImag, iLength, bInverse)
}
\arguments{
\item{pReal}{Numeric array containing the real component of the input.}

\item{pImag}{Numeric array containing the imaginary component of the input.}

\item{iLength}{An integer scalar that contains the length of the input arrays.}

\item{bInverse}{If \code{FALSE} then a forward transform is applied. If
\code{TRUE} then an inverse transform is applied.}
}
\value{
Void NIMBLE object type
}
\description{
An internal wrapper to call the fft function from the armadillo library and
expose in a way that NIMBLE can access. Is not called directly by the user
(see \code{\link{nimfft}} instead for the correct user interface).
}
\seealso{
\code{\link[nimble]{nimbleExternalCall}} \code{\link{nimfft}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
\keyword{internal}
