% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{nimfft}
\alias{nimfft}
\title{NIMBLE Interface to Apply a Forward/Inverse Fast Fourier Transform}
\usage{
nimfft(pReal, pImag, bInverse)
}
\arguments{
\item{pReal}{Numeric array containing the real component of the input.}

\item{pImag}{Numeric array containing the imaginary component of the input.}

\item{bInverse}{If \code{FALSE} then a forward transform is applied. If
\code{TRUE} then an inverse transform is applied.}
}
\value{
A matrix with 2 rows and a number of columns equal to
\code{max(length(pReal), length(pImag))}. The first row contains the real
components of the transformed output and the second row contains the
imaginary components of the transformed output.
}
\description{
Perform a forward or inverse Fast Fourier Transform using the routines
provided by the armadillo linear algebra library
(\url{https://arma.sourceforge.net/}). The function is wrapped such that the
function can be used in NIMBLE model code.
}
\seealso{
\code{\link[nimble]{nimbleFunction}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
