% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisbinom.R
\name{poisbinom_massvec}
\alias{poisbinom_massvec}
\title{Calculate Vector of Poisson Binomial Distribution Values}
\usage{
poisbinom_massvec(probVec)
}
\arguments{
\item{probVec}{A numeric vector of probability values for each of the
independant trials that comprise the Poisson binomial distribution.}
}
\value{
A numeric vector with length \code{length(probVec) + 1} containing
the probability masses from the corresponding Poisson binomial distribution.
The probability of exactly n successful trails can be found in the (n + 1)th
element of the output vector.
}
\description{
Function that provides an interface to the NIMBLE modelling framework to
create a vector of probability mass values resulting from a Poisson Binomial
distribution.
}
\details{
The calculation of the probability masses is performed according to
the discrete Fourier transform of the Poisson binomial characteristic
function as described by Hong (2013).
}
\references{
Hong (2013) On computing the distribution function for the
Poisson binomial distribution \emph{Computational Statistics and Data Analysis}
59, 41-51.
}
\seealso{
\code{\link[nimble]{nimbleExternalCall}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
