% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisbinom.R
\name{ppoisbinom}
\alias{ppoisbinom}
\title{Cumulative Probability Mass Function for the Poisson Binomial Distribution}
\usage{
ppoisbinom(q, probs, lower.tail = 1, log.p = 0)
}
\arguments{
\item{q}{Integer scalar containing the number of successful trials to
calculate the cumulative probability mass for.}

\item{probs}{Numeric vector containing the probability of success of each of
the independent trails that comprise the Poisson Binomial distribution.}

\item{lower.tail}{Logical scalar. If \code{TRUE} calculate the probability
that there \code{q} or fewer successful trials.  If \code{FALSE} calculate
the complementary cumulative distribution function.}

\item{log.p}{Logical scalar indicating whether the log of the cumulative
probability mass is required.}
}
\value{
A numeric scalar containing the cumulative probability.
}
\description{
Calculate the cumulative probability mass for the Poisson binomial
distribution. This function is encapsulated in a way that it can extend the
functionality of the NIMBLE modelling framework.
}
\details{
The calculation of the probability masses is performed according to
the discrete Fourier transform of the Poisson binomial characteristic
function as described by Hong (2013).
}
\references{
Hong (2013) On computing the distribution function for the
Poisson binomial distribution \emph{Computational Statistics and Data Analysis}
59, 41-51.
}
\seealso{
\code{\link[nimble]{nimbleExternalCall}} \code{\link{poisbinom_massvec}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
