% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisbinom.R
\name{qpoisbinom}
\alias{qpoisbinom}
\title{Quantile Function for the Poisson Binomial Distribution}
\usage{
qpoisbinom(p, probs, lower.tail = 1, log.p = 0)
}
\arguments{
\item{p}{A numeric scalar representing the cumulative probability mass for
which the corresponding quantile is going to be calculated.}

\item{probs}{Numeric vector containing the probability of success of each of
the independent trails that comprise the Poisson Binomial distribution.}

\item{lower.tail}{Logical scalar. If \code{TRUE} then the quantile is being
calculated from the lower tail of the probability distribution.  If
\code{FALSE} then the quantile is being calculated from the upper tail of the
probability distribution.}

\item{log.p}{Logical scalar indicating whether the cumulative probability
mass is given on a log scale.}
}
\value{
The number of successful trials associated with the relevant
quantile.
}
\description{
Calculate the quantiles for the Poisson binomial distribution.  This function
is encapsulated in a way that it can extend the functionality of the NIMBLE
modelling framework.
}
\details{
The calculation of the probability masses is performed according to
the discrete Fourier transform of the Poisson binomial characteristic
function as described by Hong (2013).
}
\references{
Hong (2013) On computing the distribution function for the
Poisson binomial distribution \emph{Computational Statistics and Data Analysis}
59, 41-51.
}
\seealso{
\code{\link[nimble]{nimbleExternalCall}} \code{\link{poisbinom_massvec}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
