% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisbinom.R
\name{rpoisbinom}
\alias{rpoisbinom}
\title{Random Number Generator for the Poisson Binomial Distribution}
\usage{
rpoisbinom(n, probs)
}
\arguments{
\item{n}{Integer scalar indicating the number of random samples to generate
(currently this value must be 1).}

\item{probs}{Numeric vector containing the probability of success of each of
the independent trails that comprise the Poisson Binomial distribution.}
}
\value{
An integer scalar containing a random draw of the number of
successful trials.
}
\description{
Generate a random number from the Poisson Binomial Distribution. This
function is encapsulated in a way that it can extend the functionality of the
NIMBLE modelling framework.
}
\seealso{
\code{\link[nimble]{nimbleExternalCall}}
}
\author{
Joseph D. Chipperfield, \email{joechip90@googlemail.com}
}
